pyproject.toml
[tool.poetry]
name = "airflow-cli"
version = "0.1.0"
description = ""
authors = ["Matthew Aguerreberry <matthew.aguerreberry@etermax.com>"]
readme = "README.md"
packages = [{include = "airflow_cli"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.21.1"
requests = "^2.28.2"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
ipdb = "^0.13.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
